README.md
Java Desktop Application

The application is created for Store Management purpose. It should keep record of articles on Lager that are available for market. On the other hand keep track of sold articles and full Record of Reports being able to generate it according to any filters applied to it (Date filters currently available)